complete SendPulseEmail
complete SendPulseEmail
