+++ src/app/server/services/sendPulseService.js
@@ -1,49 +1,18 @@
-import handleApiError from '@/utils/handleApiError';
-import createHttpError from 'http-errors';
-
-export const EMAIL_TEMPLATES = {
-  uk: {
-    subject: 'Дякуємо за ваше звернення!',
-    template: (name) => ({
-      html: `
-        <div style="font-family: Arial, sans-serif; padding: 20px; max-width: 600px; margin: 0 auto;">
-          <h2>Вітаємо, ${name}!</h2>
-          <p>Дякуємо за ваше звернення. Ми отримали ваше повідомлення та зв'яжемося з вами найближчим часом.</p>
-          <p>З найкращими побажаннями,<br>Команда In Harmony</p>
-        </div>
-      `,
-      text: `Вітаємо, ${name}!\n\nДякуємо за ваше звернення. Ми отримали ваше повідомлення та зв'яжемося з вами найближчим часом.\n\nЗ найкращими побажаннями,\nКоманда In Harmony`,
-    }),
-  },
-  en: {
-    subject: 'Thank you for your inquiry!',
-    template: (name) => ({
-      html: `
-        <div style="font-family: Arial, sans-serif; padding: 20px; max-width: 600px; margin: 0 auto;">
-          <h2>Hello, ${name}!</h2>
-          <p>Thank you for reaching out to us. We have received your message and will contact you soon.</p>
-          <p>Best regards,<br>In Harmony Team</p>
-        </div>
-      `,
-      text: `Hello, ${name}!\n\nThank you for reaching out to us. We have received your message and will contact you soon.\n\nBest regards,\nIn Harmony Team`,
-    }),
-  },
-};
+import { SENDPULSE_CONFIG } from '@/shared/constants/index.js'
+import createHttpError from 'http-errors'
+import { EMAIL_TEMPLATES } from '../templates/email/index.js'
 
 const getSendPulseToken = async () => {
   try {
-    const response = await fetch(
-      'https://api.sendpulse.com/oauth/access_token',
-      {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
-        body: new URLSearchParams({
-          grant_type: 'client_credentials',
-          client_id: process.env.SENDPULSE_CLIENT_ID,
-          client_secret: process.env.SENDPULSE_CLIENT_SECRET,
-        }),
-      }
-    );
+    const response = await fetch(SENDPULSE_CONFIG.URL_ACCESS_TOKEN, {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
+      body: new URLSearchParams({
+        grant_type: 'client_credentials',
+        client_id: SENDPULSE_CONFIG.CLIENT_ID,
+        client_secret: SENDPULSE_CONFIG.CLIENT_SECRET,
+      }),
+    });
 
     const data = await response.json();
     if (!data.access_token) {
@@ -55,7 +24,7 @@ const getSendPulseToken = async () => {
   }
 };
 
-const sendEmail = async ({ name, to, lang = 'uk' }) => {
+export const sendEmail = async ({ name, to, lang = 'ua' }) => {
   try {
     const token = await getSendPulseToken();
     const template = EMAIL_TEMPLATES[lang];
@@ -68,8 +37,8 @@ const sendEmail = async ({ name, to, lang = 'uk' }) => {
       email: {
         subject: template.subject,
         from: {
-          name: process.env.SENDPULSE_SENDER_NAME,
-          email: process.env.SENDPULSE_SENDER_EMAIL,
+          name: SENDPULSE_CONFIG.SENDER_NAME,
+          email: SENDPULSE_CONFIG.SENDER_EMAIL,
         },
         to: [{ name, email: to }],
         html: emailContent.html,
@@ -77,7 +46,7 @@ const sendEmail = async ({ name, to, lang = 'uk' }) => {
       },
     };
 
-    const response = await fetch('https://api.sendpulse.com/smtp/emails', {
+    const response = await fetch(SENDPULSE_CONFIG.URL_SMTP_EMAIL, {
       method: 'POST',
       headers: {
         'Content-Type': 'application/json',
@@ -98,21 +67,3 @@ const sendEmail = async ({ name, to, lang = 'uk' }) => {
     throw createHttpError(500, 'Не вдалося надіслати email');
   }
 };
-
-export default async function handler(req, res) {
-  try {
-    if (req.method !== 'POST') {
-      throw createHttpError(405, 'Метод не дозволено');
-    }
-
-    const { name, email, lang } = req.body;
-    if (!name || !email) {
-      throw createHttpError(400, "Обов'язкові поля не передані");
-    }
-
-    await sendEmail({ name, to: email, lang });
-    return res.status(200).json({ message: 'Email успішно надіслано' });
-  } catch (error) {
-    handleApiError(res, error);
-  }
-}
