+++ src/pages/api/message/telegram/index.js
@@ -1,8 +1,8 @@
-import { sendMessageTg } from '@/app/server/services/telegramService.js';
+import { sendTelegramMessage } from '@/app/server/services/telegramService.js';
 
 export default async function handler(req, res) {
   if (req.method === 'POST') {
-    return sendMessageTg(req, res);
+    return sendTelegramMessage(req, res);
   } else {
     return res.status(405).json({ message: 'Method Not Allowed' });
   }
